requirements.txt
pymupdf>=1.23.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
ftfy>=6.1.0
openai>=1.0.0
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pdfplumber>=0.10.3
tqdm>=4.66.0

spacy>=3.7.0
scispacy>=0.5.4  # Medical/scientific text processing
# Note: scispacy models need to be downloaded separately:
# pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz

